outsourced delete prompt as component
